+++ packages/superposition/src/Interface/Bdb.ts
@@ -1,3 +1,3 @@
 import { Detoxicated } from './Detoxicated';
 
-export type Bdb<T> = Error | Detoxicated<T>;
+export type Bdb<T> = Detoxicated<T> | Error;