+++ packages/unscharferelation/src/Heisenberg/__tests__/Uncertain.spec.ts
@@ -1,11 +1,6 @@
-import { MockRuntimeError } from '@jamashita/anden-error';
-import { MockValueObject } from '@jamashita/anden-object';
 import sinon, { SinonSpy } from 'sinon';
 import { UnscharferelationError } from '../../Error/UnscharferelationError';
-import { Absent } from '../Absent';
 import { Heisenberg } from '../Heisenberg';
-import { Lost } from '../Lost';
-import { Present } from '../Present';
 import { Uncertain } from '../Uncertain';
 
 describe('Uncertain', () => {
@@ -96,40 +91,6 @@ describe('Uncertain', () => {
       });
 
       expect(spy.called).toBe(false);
-    });
-  });
-
-  describe('equals', () => {
-    it('returns true if the same instance given', () => {
-      expect.assertions(1);
-
-      const heisenberg: Heisenberg<number> = Uncertain.of<number>();
-
-      expect(heisenberg.equals(heisenberg)).toBe(true);
-    });
-
-    it('returns false if the different class instance given', () => {
-      expect.assertions(1);
-
-      const heisenberg: Heisenberg<number> = Uncertain.of<number>();
-
-      expect(heisenberg.equals(new MockValueObject('mock'))).toBe(false);
-    });
-
-    it('returns true if Uncertain given', () => {
-      expect.assertions(4);
-
-      const present: Present<number> = Present.of<number>(2);
-      const absent: Absent<number> = Absent.of<number>();
-      const lost: Lost<number> = Lost.of<number>(new MockRuntimeError());
-      const uncertain: Uncertain<number> = Uncertain.of<number>();
-
-      const heisenberg: Heisenberg<number> = Uncertain.of<number>();
-
-      expect(heisenberg.equals(present)).toBe(false);
-      expect(heisenberg.equals(absent)).toBe(false);
-      expect(heisenberg.equals(lost)).toBe(false);
-      expect(heisenberg.equals(uncertain)).toBe(true);
     });
   });